Reorganized save folder/class structure
