+++ src/mod/save/data/badge/badge.cpp
@@ -1,0 +1,16 @@
+#include "save/save.h"
+
+void loadBadges(PlayerWork::Object* playerWork)
+{
+    // TODO
+}
+
+void saveBadges(PlayerWork::Object* playerWork)
+{
+    // TODO
+}
+
+void restoreBadges(PlayerWork::Object* playerWork)
+{
+    // TODO
+}
