+++ src/mod/save/data/box/box.cpp
@@ -1,0 +1,16 @@
+#include "save/save.h"
+
+void loadBoxes(PlayerWork::Object* playerWork)
+{
+    // TODO
+}
+
+void saveBoxes(PlayerWork::Object* playerWork)
+{
+    // TODO
+}
+
+void restoreBoxes(PlayerWork::Object* playerWork)
+{
+    // TODO
+}
