+++ src/mod/save/data/color_variation/color_variation.cpp
@@ -1,0 +1,16 @@
+#include "save/save.h"
+
+void loadColorVariations(PlayerWork::Object* playerWork)
+{
+    // Nothing to do
+}
+
+void saveColorVariations(PlayerWork::Object* playerWork)
+{
+    // Nothing to do
+}
+
+void restoreColorVariations(PlayerWork::Object* playerWork)
+{
+    // Nothing to do
+}
