+++ src/mod/save/data/string/string.cpp
@@ -1,0 +1,16 @@
+#include "save/save.h"
+
+void loadStrings(PlayerWork::Object* playerWork)
+{
+    // TODO
+}
+
+void saveStrings(PlayerWork::Object* playerWork)
+{
+    // TODO
+}
+
+void restoreStrings(PlayerWork::Object* playerWork)
+{
+    // TODO
+}
