+++ src/mod/save/data/trainer/trainer.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include "externals/DPData/TR_BATTLE_DATA.h"
+#include "externals/PlayerWork.h"
 
 template <int32_t size>
 struct TrainerSaveData {
@@ -29,3 +30,7 @@ struct TrainerSaveData {
         return index + GetByteCount();
     }
 };
+
+void loadTrainers(PlayerWork::Object* playerWork);
+void saveTrainers(PlayerWork::Object* playerWork);
+void restoreTrainers(PlayerWork::Object* playerWork);
