+++ src/mod/save/migration/save_migration.h
@@ -1,0 +1,7 @@
+#include "externals/PlayerWork.h"
+
+// Main function to migrate the save file.
+void migrate(PlayerWork::Object* playerWork);
+
+// Migrates from Vanilla to Release 3.0.
+void migrateFromVanilla(PlayerWork::Object* playerWork);
